+++ docs/src/markdown/about/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.5.3
+
+- **FIX**: Previously, all pseudo classes' selector lists were evaluated as one big group, but now each pseudo classes' selector lists are evaluated separately.
+- **FIX**: CSS selector tokens are not case sensitive.
+
 ## 0.5.2
 
 - **FIX**: Add missing `s` flag to attribute selector for forced case sensitivity of attribute values.